construction-automation/app4-site-report/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.0
openai==1.45.0
python-docx==1.1.2
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
